kulfy-agent/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn==0.27.0

# LangGraph and LangChain (compatible versions)
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# OpenAI
openai>=1.10.0

# Web scraping
beautifulsoup4==4.12.3
requests==2.31.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp==3.9.1
Pillow>=10.3.0
